+++ src/pages/index.ts
@@ -11,5 +11,5 @@ export { default as VerifyEmail } from "./user/findId/verification/VerifyEmail";
 export { default as ShowId } from "./user/findId/ShowId";
 export { default as validateId } from "../components/validation/id-validation";
 export { default as validatePw } from "../components/validation/pw-validation";
-export { default as TransactionHistory } from "./TransactionHistories/TransactionHistory";
+export { default as Transaction } from "./Transaction";
 export { default as Register } from "./user/register/Regstier";
